Utilize Custom Network selector in Apache App
A new control allowing to select existing networks for the instance to
join has been implemented in Ie256c36523c67f27e4cb0b7c9aeed76b2abd28af

This patch modifies Apache HTTP Server application to utilize this
control. It adds it to the "configure instance" dialog step of the
wizard and passes the values returned from the control to the 'networks'
section of instance object in the generated Object Model.

Note, that this also bumps the "Version" of the used UI definition
notaion to 2.1, so this app becomes incompatible with pre-Liberty Murano
releases and with the Murano built from Liberty master prior the landing
of Ie256c36523c67f27e4cb0b7c9aeed76b2abd28af.

Depends-On: Ie256c36523c67f27e4cb0b7c9aeed76b2abd28af

Change-Id: I98ec3ce1992cc6ddbabd3f9ea79891dc952c51a6