Fixed an issue occasionally causing an empty config for etcd slave nodes
The script which adds new nodes to etcd cluster now does up to 30 attempts
and fails (returns a non-zero exit code) if none are successful.

A minor performance optimization has been done as well, so no unneeded
waits happen when the command is successful and no temporary files are used.

Closes-Bug: #1494997
Change-Id: Ic0552f388518c119925da902b64797bbf96f979f