[Ops Manager] Unify configuration for different versions of app
Previously, app exported file with full list of settings to Ops
Manager. But some values of installation settings are changing from
release to release and it's better to use bare settings of already
deployed but non-configured Ops Manager and merge them with data
specified by user which are the same for all versions. After merging
ready file can be exported to OpsManager.
This patch implemets this idea

Change-Id: I181dfe1b5183a224f206784a46587725aec38cd9